refactor: extract shared @TestConfiguration for JPA auditing into TestJpaAuditingConfig (#53)
