+++ src/test/java/com/alimmit/golf/TestJpaAuditingConfig.java
@@ -1,0 +1,31 @@
+package com.alimmit.golf;
+
+import java.util.Optional;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.data.repository.query.SecurityEvaluationContextExtension;
+
+@TestConfiguration
+@EnableJpaAuditing
+public class TestJpaAuditingConfig {
+
+  @Bean
+  AuditorAware<String> auditorAware() {
+    return () -> {
+      Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+      if (authentication == null || !authentication.isAuthenticated()) {
+        return Optional.empty();
+      }
+      return Optional.of(authentication.getName());
+    };
+  }
+
+  @Bean
+  SecurityEvaluationContextExtension securityEvaluationContextExtension() {
+    return new SecurityEvaluationContextExtension();
+  }
+}
